requirement.txt
transformers
torch
fastapi
streamlit
Pillow
numpy
uvicorn